fix: PR Refactor
merged docs to the example README
merged both contorl panel to be in the same one
added comments and docs
