+++ examples/geometry/index.html
@@ -5,7 +5,7 @@
     <meta
       name="viewport"
       content="width=device-width, initial-scale=1.0, user-scalable=no" />
-    <title>Example:</title>
+    <title>Example: Map with Geometry</title>
 
     <style>
       body {
