+++ examples/geometry/src/encoded-polygon-data.ts
@@ -1,3 +1,4 @@
+// Example of encoded polygons of certain zipcode areas near the border of Nebraska and Iowa.
 export const POLYGONS = [
   '}iyzFjf`hQv@mIpCf@ZDbG@jCF^Il@HAnItDDhJBBoGrE@UrAGz@E`Cl\\Hl@CBmMhE@v@eAN@bBs@Aw@DUJSBKK@@WhCuD~CuF`AuAzEwFzw@w}@]mCvGyEcAuIfCfKfBwBbCxFp@nBNO`@r@tGiLzR_[ZL|AVvDg@dWr\\l@p@TDjAfCZz@b@dB\\~BFz@FdB@pIlAeBfAmBzS{Yr@{@fAqAzP}Uo@kHAUdCEfAK\\@|A\\rCBdDX`APjAj@VVXLb@b@t@b@@JnA|@nBdA`@b@ZNT\\h@Z^`@r@b@d@h@|AjAJPF@jAhArAt@NPJBHHddARrw@e@tH@b@f@xB|JbDpH^zE\\^Xx@`CnD~AdBtAbAtBbD~@~CFx@V`AHpB^dCWTeAZ{El@eAj@eAbAuHxBiAz@iAfBkAfAgA|@{Av@_A|@yBjDaD|D{DtH_BtE_AdB[jAQpAeBdF]zCuApEUhCa@xAq@fDItCcAfFGhCYbB_@pF_@hB[hCGtA@|Aq@vFFvCQtFBfB[fGPtCCfFLpBIjFVtCCxHLt@LbCFlFCjDRlBBfETbBH|DRbBBpBpDj]d@zBPvBr@zETxE^pCn@pHlAlQnBbXC`DZdDHlCCjAQzAH|ATtMJpBEjD^jKQxEHxHOlB_@rBCzBDfBKbCUbBo@hBM|AEvCYhCq@rBkAbB_CtIgAbCaAdAuAr@kD~FqBpCi@`@iAb@yBdCq@d@}@\\}BxBwBz@oC|Bu@\\}@PsCrBqBViB|@sAfAeBN{D|A_AHoAEsFnAuALy@`@_ALgAGwAg@k@?cCf@cDZaAEeAUiINcBEaB[wA_A}Ls@sAOyAu@q@O_AA_AQ{@_@u@o@qCk@{@c@s@s@eBi@sBkAyA{BeF{CcBs@{AkAo@w@g@cAeCoBcPcT{DmFqAy@wEmFyAy@cGyFqCuAoAc@uDiCoAi@sASyBy@eEqBaC}@cDYgC}@uA?_C[yBo@kBFkBImALgCCw@Q}APmFAqHp@u@?yA`@_E`@mBbAmEj@kBhAqCp@}BzAiA`@q@L{D|CeBf@gAn@oClCqBt@eApAw@j@_Bf@_CfBwA^iDbBqJ~DgFrCmEvAoAN_IdDeEt@eH`CyEn@}E`@iE`AaA@}@McEr@eD?}CVcDEyGiCoAKcA]uD{DqBwAw@qA{AaEoDcEe@y@}@_Ds@_AgAwB_@uAUcBuA{D_AgFQeBiAaGUkC[wH]oD?aDyAgOcA{Me@{AgBoIaAwGe@qAq@oAk@wCYwCy@gC{GwZiGcTgGoR{BkFmDuJmByD{DeIeA_BqBwDeF}G{A{AsAy@sAuBgAmAq@e@}@YkAeAiBw@qEmCUEDo]AkAHgSNoAEgNDmGpPh@gAOrBFjACjf@|AdEHx@?tCC|e@aArWA?~Hd@Wb@[pCkCD}B`C?Ll@HC@k@rE?[Rx@CvEH^B?StDb@VIbBj@@Rb@N@wNBqCH[?kJZiC`@qEFuABiAf@yEvNjC',
   'enxzFjh`hQuDE@oIm@I_@HkCGcGA[EqCg@w@lIwNkCg@xEChAGtAa@pE[hC?jJIZCpCAvNc@OAScBk@WHuDc@?R_@CwEIy@BZSsE?Aj@IBMm@aC?E|BqCjCc@Ze@V?_IsW@}e@`AuCBy@?eEIkf@}AkABsBGfANqPi@ElGDfNOnAAhCcyAbSkKxAmDn@aEdAiEzAqDbBuC`BsClBwB`BgA~@KkEKoB?s@D}@H]J[|AiDX{@z@uDe@{@H?`@KHMFS@G@eCN_A@w@g@cDKyBHkAAe@k@mCWmBQ}@Qs@a@}@Gw@Mk@Qi@]s@k@q@YSGMI}AE[Mk@IKGC}AQSIg@e@m@}@{@_BoAmBGc@GQKK}@inAe@@mAE[Du@XeAAqAf@q@Hi@TUVs@BWNQZSj@DmAC_@PICa@BgBAiDBqBC_@h@oMcAwJL{AuAeAi@[]KsAS{RmB{B[qGuBiFmBqB}@g@QcAW{AQgACaBDqBL_AD]AwDo@gBUoAIy@@aANWjh@{J?sjCy@{CE\\G\\K\\O`As@jGoFXa@Rg@N}@BcULw@Pc@V]pAeA`CgBrBsBrEeFdAoAfF{HVa@p@gBnA_EVc@PQZQTEhBMbAYv@YNKLMR[Jk@HiIHs@Pa@T[vAgAbEoCh@e@z@}@xDiFZo@La@XsA`@sAbBwDl@aBNy@Bc@?s@QcCEeAEkNGaEQ_AUu@qA}CKk@Em@AcCByOCcBsDFwSHeMCwCEBkHAiSFug@DyQjLGb@eCt@cD|BaJdL_@jHem@hDo@XQVGf@A^O`@]nAm@l@MRKAIJKrG{@f@O\\i@NOr@Ct@D\\X^O`A_BNXNSHHRMT[R_@bAyDTMPm@^A`@_@M_@x@c@P[l@i@tA_A|Aa@l@o@LUHYb@s@lDkDp@uCTY`@a@JWNcAEmAm@c@E[Xs@Z[~@{A^_@Lo@?i@PUvCuAj@w@`@qACs]EaRTh@Vb@j@r@f@b@v@b@d@PVFp@H`ZGb^FlD?dGId@Gd@M\\O\\Uh@i@f@u@rAeCg|@yaAmCoCiHyGeAeAoLsMM_ADQMkDTmVqAkOo@{y@D[@iBCwt@@uH?cIIq@Eg~@xB@nHA~]Dl[HhL?~P@xGDdCNpJFPB`@NZ\\Pf@Hj@DbEJ|^H|Od@CnVBbU?q@wCGq@Iy]CiU@}KHkl@FcRC{EDgJ@kWlB?zHD|SAxDDxCAdEBnGCrSHdAC~@GhFu@WsEIwE?oQIye@a@_n@bA@bHC|NB`OA|@CnABzSBf_@@vy@F@|AlCgC~BmA^GLQ~@EjBo@Zu@dHb@lO\\bZb@~ST`KVfE\\|CB?rL@nFCxIAnc@?px@CtTCpD~D@zSGpLBImW@sCHYFGHC|GBza@AAwHJkb@Fuo@jCTfF\\vK`BfAVzAP~@PhJt@bd@Hr[EtAb@v@DjG~Ah@DLPf@DlHfBlAPVPxD`AhVpFzDdBpCh@v@PTJzc@\\pOQzDSpFj@lFjBbCxAzUbRhBfAhAb@x@PhCDzCn@tTfD~ARvTfDvKpBbDt@hg@|IzG`At@VhI`BpAPt@P~Ez@l@\\Z?zKbCbCn@h@FDlcAOhaA?fg@A`BKzAXth@Stn@AhhAL~^ApkAQhGDtw@L?ZzGAdCRbcA@tM`E_AnLqCh@It@CzBD|@@t@GfAUt@]|BcBjHiGtA_BdCqDh@i@~@g@h@QfB[~@WhAe@t@_@v@m@pGcGl@e@lCgBl@pCd@jDxAfA~CdBdCxC^p@dj@Xl@poAcl@rCP|pA|BjAnEtBpAd@|IlB|H|@tBNfFHtC?zBC~CKh@FdAVb@LHL@TChDay@aBsAMmAU@fBoa@w@@~MAxN@hJ~Ls@pFY|DYzLm@xWwApOw@lEExDBxDDfEPzEXvD^zDd@tEv@tDt@j}At^ncA|UpGvApDn@lC\\jBRxCR~BJbFBjDEdBEvCQlCWrAQbDi@lCg@`Ck@td@oM|JkC|g@uNhKqCGzv@BT@fAAri@@r@BXFHj@FHLKvSe@v{ANvpAQnH{DeGoDaEMY{CcFaRiTsFwEsHqEmBm@aGmCcHiB{FkA}TeBsDPkB?uCLmCf@yBr@}AHuAh@oAEsDfAyIvCoGvAeBn@wDf@aGxAeM`BkHxAgTpAcASoAXoAA{Gw@cYeB{OVcHj@iGvAqDZuBbAgJbDk@^oBZcBv@kAXaBnA_DhBq@JwD|B}CxA{DpDm@fAq@r@gAn@e@h@}BbAcAvAuC~B_Ah@eEnDm@l@Ub@eBbAuCbCyBtAa@j@_B~@y@x@sChAwD`CaAfA}BfAkAXgCtAs@r@{CfA{APyGnCoFt@eEx@{@^sAVi@@o@UaADwB^iA^yBHkBOuIp@uAMiAe@_DVwBKsA]sDIaB{@cDU{@[}@o@wAg@}G_B_BqAcL}E_A{@kC_BeKiJmGyHs@oAiCgEuAkCgAeE}@_C_A_E]iBm@wG[uJDsGt@qHfAaFf@wDjA_Hl@uB\\sErAcIb@cDrA_T_@aO_@mEYaBI}Ak@_EiAcEcDkHcDwEgBaBaAwA{DmC}@UkEuBy@CcCm@kDYuCG{DJwBKmAP_@eCIqBWaAGy@_A_DuBcDuAcA_BeBaCoDYy@]_@_@{EcDqHyB}Jc@g@uHAsw@d@edASIIKCOQsAu@kAiAGAKQ}AkAe@i@s@c@_@a@i@[U][Oa@c@oBeAoA}@AKu@c@c@c@YMWWkAk@aAQeDYsCC}A]]AgAJeCD@Tn@jH{P|UgApAs@z@{SzYgAlBmAdBAqIGeBU}BO}@c@eB[{@kAgCUEm@q@eWs\\wDf@}AW[M{R~ZuGhLa@s@ONq@oBcCyFgBvBgCgKbAtIwGxE\\lC{w@v}@{EvFaAtA_DtFiCtDAVJACJKRET@v@cBr@OAw@dAiEAClMm@Bm\\IDaCF{@TsAsEACnGiJC',
